+++ lib/mongo/mongo_db_connection.ex
@@ -95,15 +95,17 @@ defmodule Mongo.MongoDBConnection do
     end
   end
 
-  defp tcp_connect(opts, s) do
-    host      = (opts[:hostname] || "localhost") |> to_charlist
-    port      = opts[:port] || 27017
-    sock_opts = [:binary, active: false, packet: :raw, nodelay: true]
-                ++ (opts[:socket_options] || [])
+  defp tcp_connect(opts, state) do
 
-    s = Map.put(s, :host, "#{host}:#{port}")
+    {host, port} = Utils.hostname_port(opts)
+    sock_opts   = [:binary, active: false, packet: :raw, nodelay: true] ++ (opts[:socket_options] || [])
 
-    case :gen_tcp.connect(host, port, sock_opts, s.connect_timeout) do
+    s = case host do
+      {:local, socket} -> Map.put(state, :host, socket)
+      hostname         -> Map.put(state, :host, "#{hostname}:#{port}")
+    end
+
+    case :gen_tcp.connect(host, port, sock_opts, state.connect_timeout) do
       {:ok, socket} ->
         # A suitable :buffer is only set if :recbuf is included in
         # :socket_options.
@@ -113,7 +115,7 @@ defmodule Mongo.MongoDBConnection do
 
         {:ok, %{s | connection: {:gen_tcp, socket}}}
 
-      {:error, reason} -> {:error, Mongo.Error.exception(tag: :tcp, action: "connect", reason: reason, host: s.host)}
+      {:error, reason} -> {:error, Mongo.Error.exception(tag: :tcp, action: "connect", reason: reason, host: state.host)}
     end
   end